+++ quantforge/__init__.py
@@ -768,6 +768,7 @@ from .signal_features import (
     spectral_flatness,
 )
 from .interval import Interval
+from .bernoulli import bernoulli_number, faulhaber, bernoulli_sequence
 from .levenberg import levenberg_marquardt
 from .ewma_cov import ewma_covariance_matrix, ewma_correlation_matrix
 from .vol_convert import black_to_normal_vol, normal_to_black_vol
@@ -1158,7 +1159,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "3.22.0"
+__version__ = "3.23.0"
 
 __all__ = [
     "OptionType",
@@ -2202,6 +2203,9 @@ __all__ = [
     "spectral_bandwidth",
     "spectral_flatness",
     "Interval",
+    "bernoulli_number",
+    "faulhaber",
+    "bernoulli_sequence",
     "levenberg_marquardt",
     "ewma_covariance_matrix",
     "ewma_correlation_matrix",